Dynamic Recognition, Changed Net, Removed unused buttons
Dynamic Recognition is performed as user is drawing a digit.

Changed the net to one with 3 hidden layers. Current layout is
784-70-50-20-10 (numbers of neurons in each layer)

Removed unused buttons - this includes Invert and Recognize, which is
not needed after implementing dynamic recognition.

Additionally, a drawWeights function was implemented to be able to
display how each pixel affects decision in a simple 2 layer net.