Updated style and UI
UI updates include adding explanation of how the app workos as well as
controls and creting a keybinding to clear the canvas. The canvas itself
was enlarged to provide more are for drawing and showcasing that digits
can be drawn anywhere and still be recognized.

Cleaned up commented out code that was not used.